refactor(daemon): extract health, auth, and dashboard routes from daemon.ts
