+++ package.json
@@ -67,5 +67,8 @@
     "bun",
     "sqlite",
     "signet"
-  ]
+  ],
+  "dependencies": {
+    "@types/bun": "^1.3.12"
+  }
 }
